+++ src/types/UserProjectData.ts
@@ -3,6 +3,7 @@ import { ProjectResource } from "./ProjectResource";
 export interface UserProjectData {
     org : string,
     name : string,
+    owner? : string,
     guidelines : string,
     resources : ProjectResource[],
 }